+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ## [Unreleased]
 ### Changed
 - Divide Editor#move event into Editor#move and Editor#enter events.
+- Prefer underscore over hyphen as file name.
 
 ## [0.4.0]
 ### Added