+++ CHANGELOG.md
@@ -7,11 +7,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
 ## [Unreleased]
+
+## [0.5.0] - 2016-03-14
 ### Changed
 - Divide Editor#move event into Editor#move and Editor#enter events.
 - Prefer underscore over hyphen as file name.
 
-## [0.4.0]
+## [0.4.0] - 2016-03-14
 ### Added
 - Enable to preload third party editor classes via `Textcomplete.editors`.
 - Enable to select dropdown by tab key.
@@ -23,7 +25,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ### Fixed
 - Fix dropdown position when window is scrolled.
 
-## [0.3.0]
+## [0.3.0] - 2016-03-10
 ### Added
 - Add "Getting Started", "Development" and "Events" documents.
 - Add a contributing guide.
@@ -74,7 +76,8 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/yuku-t/textcomplete/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/yuku-t/textcomplete/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/yuku-t/textcomplete/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/yuku-t/textcomplete/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/yuku-t/textcomplete/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/yuku-t/textcomplete/compare/v0.1.2...v0.2.0